+++ exclusive_access_off.go
@@ -1,0 +1,5 @@
+//go:build !flecs_exclusive_access
+
+package flecs
+
+const flecsExclusiveAccess = false //nolint:unused