+++ exclusive_access_on.go
@@ -1,0 +1,5 @@
+//go:build flecs_exclusive_access
+
+package flecs
+
+const flecsExclusiveAccess = true //nolint:unused